+++ CHANGELOG.md
@@ -1,5 +1,11 @@
 # Changelog
 
+### Unreleased
+
+- Deprecate `earthaccess.get_s3fs_session` and `Store.get_s3fs_session`. Use
+  `earthaccess.get_s3_filesystem` and `Store.get_s3_filesystem`, respectively,
+  instead ([#766](https://github.com/nsidc/earthaccess/issues/766))([**@Sherwin-14**](https://github.com/Sherwin-14))
+
 ## [v0.10.0] 2024-07-19
 
 ### Changed
@@ -15,6 +21,7 @@
   ([#584](https://github.com/nsidc/earthaccess/pull/584))
   ([**@danielfromearth**](https://github.com/danielfromearth),[**@chuckwondo**](https://github.com/chuckwondo),
    [**@jhkennedy**](https://github.com/jhkennedy),[**@mfisher87**](https://github.com/mfisher87))
+
 
 ### Added